feat(renderer): add function for preparing single field props
