+++ packages/react-form-renderer/src/prepare-component-props/index.d.ts
@@ -1,0 +1,1 @@
+export { default } from './prepare-component-props';
