+++ packages/react-form-renderer/src/prepare-component-props/index.js
@@ -1,0 +1,1 @@
+export { default } from './prepare-component-props';
