+++ packages/react-form-renderer/src/prepare-component-props/prepare-component-props.d.ts
@@ -1,0 +1,13 @@
+import { AnyObject, ComponentMapper } from "../common-types";
+import { ActionMapper } from "../form-renderer";
+
+export interface PrepareComponentPropsOptions {
+    component: string;
+    rest: AnyObject;
+    componentMapper: ComponentMapper;
+    actionMapper: ActionMapper;
+}
+
+declare function prepareComponentProps(PrepareComponentPropsOptions: AnyObject): AnyObject;
+
+export default prepareComponentProps;
